Add issues: read permissions to docs
Fixes #1710

Starting from digger v0.6.41, at least `issues: read` permission is
required to check if PR number is an issue or not. It should be clearly
stated in the document so that other folks are not confused.

I'm not sure when `issues:write` is needed, please add another note
where appropriate.